requirements.txt
# PyTorch con soporte CUDA 11.8
torch
torchvision
torchaudio

# MS-COCO API
pycocotools

# Métricas de evaluación
torchmetrics

# Procesamiento de lenguaje natural
nltk

# Procesamiento de imágenes
pillow

# Visualización
matplotlib

# Utilidades
numpy
tqdm
  
